add workのコレクションページ作成
