+++ src/css/block/_index.scss
@@ -2,4 +2,5 @@
 @forward "./footer";
 @forward "./mv";
 @forward "./about";
+@forward "./work/works";
 @forward "./404";
