+++ src/css/layout/_index.scss
@@ -4,3 +4,4 @@
 @forward "./content";
 @forward "./mv";
 @forward "./about";
+@forward "./work/works";
